+++ src/types.ts
@@ -145,6 +145,9 @@ export type AppSettingsAPIResponse<StreamChatGenerics extends ExtendableGenerics
     };
     revoke_tokens_issued_before?: string | null;
     search_backend?: 'disabled' | 'elasticsearch' | 'postgres';
+    sns_key?: string;
+    sns_secret?: string;
+    sns_topic_arn?: string;
     sqs_key?: string;
     sqs_secret?: string;
     sqs_url?: string;
@@ -340,6 +343,12 @@ export type CheckPushResponse = APIResponse & {
 };
 
 export type CheckSQSResponse = APIResponse & {
+  status: string;
+  data?: {};
+  error?: string;
+};
+
+export type CheckSNSResponse = APIResponse & {
   status: string;
   data?: {};
   error?: string;
@@ -1646,6 +1655,9 @@ export type AppSettings = {
   };
   reminders_interval?: number;
   revoke_tokens_issued_before?: string | null;
+  sns_key?: string;
+  sns_secret?: string;
+  sns_topic_arn?: string;
   sqs_key?: string;
   sqs_secret?: string;
   sqs_url?: string;
@@ -2270,6 +2282,12 @@ export type TestSQSDataInput = {
   sqs_key?: string;
   sqs_secret?: string;
   sqs_url?: string;
+};
+
+export type TestSNSDataInput = {
+  sns_key?: string;
+  sns_secret?: string;
+  sns_topic_arn?: string;
 };
 
 export type TokenOrProvider = null | string | TokenProvider | undefined;