+++ test/unit/channel.js
@@ -157,6 +157,39 @@ describe('Channel count unread', function () {
 
 		expect(channel.countUnreadMentions()).to.be.equal(1);
 	});
+
+	describe('channel.lastRead', () => {
+		let channelResponse;
+		beforeEach(() => {
+			channelResponse = generateChannel();
+			channel = client.channel(channelResponse.channel.type, channelResponse.channel.id);
+			channel.initialized = true;
+		});
+
+		it('should return null if no last read message', () => {
+			expect(channel.lastRead()).to.eq(null);
+		});
+
+		it('should return last read message date', () => {
+			const last_read = new Date();
+			const messages = [generateMsg()];
+			channel.state.read[user.id] = {
+				last_read,
+				last_read_message_id: messages[0].id,
+				user: user,
+				unread_messages: 0,
+			};
+			channel.state.addMessagesSorted(messages);
+			expect(channel.lastRead()).to.eq(last_read);
+		});
+
+		it('should return undefined if client user is not set (server-side client)', () => {
+			client = new StreamChat('apiKey', 'secret');
+			channel = client.channel(channelResponse.channel.type, channelResponse.channel.id);
+			channel.initialized = true;
+			expect(channel.lastRead()).to.be.undefined;
+		});
+	});
 });
 
 describe('Channel _handleChannelEvent', function () {
@@ -496,6 +529,30 @@ describe('Channel _handleChannelEvent', function () {
 			expect(channel.state.members[user.id].banned).eq(expectAfterSecond.banned);
 			expect(channel.state.members[user.id].shadow_banned).eq(expectAfterSecond.shadow_banned);
 		});
+	});
+});
+
+describe('Uninitialized Channel', () => {
+	const user = { id: 'user' };
+	let client;
+	let channel;
+
+	beforeEach(() => {
+		client = new StreamChat('apiKey');
+		client.user = user;
+		client.userID = user.id;
+		client.userMuteStatus = (targetId) => targetId.startsWith('mute');
+		channel = client.channel('messaging', 'id');
+		channel.initialized = false;
+		channel.offlineMode = false;
+	});
+
+	it('returns 0 mentions in unread messages', () => {
+		expect(channel.countUnreadMentions()).to.eq(0);
+	});
+
+	it('reports no lastRead data', () => {
+		expect(channel.lastRead()).to.eq(null);
 	});
 });